@@ -36,12 +36,13 @@
     "@natlibfi/express-validate-content-type": "^1.0.1",
     "@natlibfi/identifier-services-commons": "git+https://github.com/natlibfi/identifier-services-commons.git#feature/apiError",
     "@natlibfi/melinda-commons": "^7.0.0",
-    "@natlibfi/passport-atlassian-crowd": "^1.0.0",
+    "@natlibfi/passport-atlassian-crowd": "^2.0.2",
     "ajv": "^6.10.2",
     "body-parser": "^1.19.0",
     "chai-http": "^4.3.0",
     "cors": "^2.8.5",
     "express": "^4.17.1",
+    "jwt-token-encrypt": "^1.0.4",
     "moment": "^2.24.0",
     "mongodb": "^3.3.0",
     "path": "^0.12.7"
@@ -57,17 +58,18 @@
     "babel-eslint": "^10.0.2",
     "babel-plugin-istanbul": "^5.1.4",
     "babel-plugin-rewire": "^1.2.0",
-    "cross-env": "^5.2.0",
-    "eslint": "^5.9.0",
+    "cross-env": "^6.0.3",
+    "eslint": "^6.5.1",
     "eslint-config-xo": ">=0.25.0 <1.0.0",
     "eslint-plugin-import": "^2.18.0",
-    "eslint-plugin-mocha": "^5.3.0",
+    "eslint-plugin-mocha": "^6.2.0",
+    "fs": "0.0.2",
     "istanbul": ">=0.4.5 <1.0.0",
     "mocha": "^6.0.2",
-    "nock": "^10.0.3",
+    "nock": "^11.5.0",
     "nodemon": "^1.19.1",
     "npm-check": "^5.9.0",
-    "nyc": "^13.1.0"
+    "nyc": "^14.1.1"
   },
   "eslintConfig": {
     "parser": "babel-eslint",
