@@ -5,13 +5,20 @@ export const API_URL = readEnvironmentVariable('API_URL', {
 	defaultValue: 'http://localhost:8080'
 });
 
+export const SMTP_URL = readEnvironmentVariable('SMTP_URL');
+export const API_EMAIL = readEnvironmentVariable('API_EMAIL');
+
+export const API_CLIENT_USER_AGENT = readEnvironmentVariable('API_CLIENT_USER_AGENT', {defaultValue: '_RECORD-IMPORT-CONTROLLER'});
+export const API_USERNAME = readEnvironmentVariable('API_USERNAME');
+export const API_PASSWORD = readEnvironmentVariable('API_PASSWORD');
+
 export const whiteList = readEnvironmentVariable('CORS_WHITELIST', {
 	defaultValue: ['http://localhost:8080'],
 	format: JSON.parse
 });
 
 export const QUERY_LIMIT = readEnvironmentVariable('QUERY_LIMIT', {
-	defaultValue: 10
+	defaultValue: 5, format: v => Number(v)
 });
 
 export const HTTP_PORT = readEnvironmentVariable('HTTP_PORT', {
@@ -23,6 +30,10 @@ export const MONGO_URI = readEnvironmentVariable('MONGO_URI', {
 	defaultValue: 'mongodb://localhost:27017/db'
 });
 export const PASSPORT_LOCAL_USERS = readEnvironmentVariable('PASSPORT_LOCAL_USERS');
+export const PASSPORT_LOCAL = readEnvironmentVariable('PASSPORT_LOCAL');
+
+export const PRIVATE_KEY_URL = readEnvironmentVariable('PRIVATE_KEY_URL');
+export const UI_URL = readEnvironmentVariable('UI_URL', {defaultValue: 'http://localhost:8080'});
 
 export const MONGO_DEBUG = readEnvironmentVariable('MONGO_DEBUG', {
 	defaultValue: false,
