@@ -1,4 +1,3 @@
-/* eslint-disable no-shadow-restricted-names */
 /**
  *
  * @licstart  The following is the entire license notice for the JavaScript code in this file.
@@ -226,7 +225,7 @@ export default function (collectionName, collectionContent) {
 
 	function getValidator(schemaName) {
 		const str = readFileSync(path.join(__dirname, '..', 'api.json'), 'utf8')
-			.replace(new RegExp('#/components/schemas', 'gm'), 'defs#/definitions');
+			.replace(/#\/components\/schemas/gm, 'defs#/definitions');
 
 		const obj = JSON.parse(str);
 
