@@ -29,8 +29,9 @@
 import HttpStatus from 'http-status';
 import {ApiError} from '@natlibfi/identifier-services-commons';
 
-import {hasAdminPermission, hasSystemPermission, hasPublisherAdminPermission} from './utils';
+import {hasAdminPermission, hasSystemPermission, hasPublisherAdminPermission, createLinkAndSendEmail, local} from './utils';
 import interfaceFactory from './interfaceModules';
+import {PASSPORT_LOCAL, PRIVATE_KEY_URL} from '../config';
 
 const userInterface = interfaceFactory('userMetadata', 'UserContent');
 
@@ -46,7 +47,10 @@ export default function () {
 
 	async function create(db, doc, user) {
 		if (hasAdminPermission(user)) {
-			const result = await userInterface.create(db, doc, user);
+			const {localUser} = local();
+			await localUser.create({PASSPORT_LOCAL: PASSPORT_LOCAL, doc: doc});
+			const newDoc = {...doc, id: doc.email};
+			const result = await userInterface.create(db, newDoc, user);
 			return result;
 		}
 
@@ -55,8 +59,10 @@ export default function () {
 
 	async function read(db, id, user) {
 		const result = await userInterface.read(db, id);
+		const {localUser} = local();
+		const localResult = await localUser.read({PASSPORT_LOCAL: PASSPORT_LOCAL, email: result.email});
 		if (hasAdminPermission(user) || (hasPublisherAdminPermission(user) && result.publisher === user.id)) {
-			return result;
+			return {...result, ...localResult};
 		}
 
 		throw new ApiError(HttpStatus.FORBIDDEN);
@@ -80,12 +86,22 @@ export default function () {
 		throw new ApiError(HttpStatus.FORBIDDEN);
 	}
 
-	async function changePwd(user) {
-		if (hasAdminPermission(user) || hasSystemPermission(user)) {
-			return null;
-		}
+	async function changePwd(doc, user) {
+		if (doc.newPassword) {
+			if (hasAdminPermission(user) || hasSystemPermission(user)) {
+				const {localUser} = local();
+				return localUser.update({PASSPORT_LOCAL: PASSPORT_LOCAL, user: doc});
+			}
 
-		throw new ApiError(HttpStatus.FORBIDDEN);
+			throw new ApiError(HttpStatus.FORBIDDEN);
+		} else {
+			const result = await createLinkAndSendEmail({request: doc, PRIVATE_KEY_URL: PRIVATE_KEY_URL, PASSPORT_LOCAL: PASSPORT_LOCAL});
+			if (result !== undefined && result.status === 404) {
+				throw new ApiError(HttpStatus.NOT_FOUND);
+			}
+
+			return result;
+		}
 	}
 
 	async function query(db, {queries, offset}, user) {
@@ -101,4 +117,3 @@ export default function () {
 		throw new ApiError(HttpStatus.FORBIDDEN);
 	}
 }
-
